Invert inline-policy / policy dependency to honour §6 seam
The shared inline-policy module imported from triage/policy/, which
broke the deprecatability seam: removing the policy mode would leave
the inline-policy boundary broken even though it is meant to be
reusable by any future inline-policy caller.

Move the shared pieces — wire enum literals, GraphQL enum-name
mapping, and the byte encoder — into triage/inline-policy/ so they
have zero imports from triage/policy/. The encoder now takes a local
minimal PacketAttrRule wire shape rather than the storage PacketAttr
type. The storage→wire translator that consumes TriagePolicyRow
moves back into triage/policy/inline-translator.ts where it
legitimately depends on the storage namespace. Re-exports in
triage/policy/types.ts keep existing Zod schemas compiling against
the same identifiers.

Also tighten the PolicyTriageScoreSnapshot doc comment: every
persisted row carries at least one matching policy score; no-match
periods materialise as ready runs with zero policy_triaged_event
rows, matching the Round 3 runner behaviour.

Part of #460
