+++ src/lib/triage/inline-policy/index.ts
@@ -1,10 +1,15 @@
 /**
- * Inline-policy boundary barrel (1B-6 / discussion #447 §3.5).
+ * Inline-policy seam (1B-6 / discussion #447 §3.5).
  *
- * Bridges the stored TriagePolicy shape (#459, JSONB) to the
- * `eventListWithTriage` `EventTriagePolicyInput` shape on the wire.
- * Shared with any future inline-policy caller; not part of the
- * policy/ deprecatability namespace.
+ * Owns the wire enums, GraphQL enum-name mapping, and the byte-array
+ * encoder for review-web's `eventListWithTriage` `EventTriagePolicyInput`.
+ * Shared with every inline-policy caller; deliberately NOT part of the
+ * `triage/policy/` deprecatability namespace, so removing the policy
+ * mode leaves the seam in place for other callers.
+ *
+ * The storage→wire translator that consumes `TriagePolicyRow` lives at
+ * `src/lib/triage/policy/inline-translator.ts` — re-exported there
+ * because it legitimately depends on the storage shape.
  */
 
 export {
@@ -12,10 +17,25 @@ export {
   encodeRuleBytes,
   encodeValueByKind,
   InlinePolicyEncodingError,
+  type PacketAttrRule,
 } from "./encode";
 export {
-  type EncodedConfidenceInput,
-  type EncodedEventTriagePolicyInput,
-  type EncodedResponseInput,
-  translatePolicyToInlineInput,
-} from "./translate";
+  cmpKindToGraphql,
+  rawEventKindToGraphql,
+  responseKindToGraphql,
+  threatCategoryToGraphql,
+  valueKindToGraphql,
+} from "./graphql-names";
+export {
+  CMP_KINDS,
+  type CmpKind,
+  RANGE_CMP_KINDS,
+  RAW_EVENT_KINDS,
+  type RawEventKind,
+  RESPONSE_KINDS,
+  type ResponseKind,
+  THREAT_CATEGORIES,
+  type ThreatCategory,
+  VALUE_KINDS,
+  type ValueKind,
+} from "./kinds";
